requirements.txt
attrs==20.3.0
aws-cdk.assets==1.23.0
aws-cdk.aws-apigateway==1.23.0
aws-cdk.aws-applicationautoscaling==1.23.0
aws-cdk.aws-autoscaling-common==1.23.0
aws-cdk.aws-certificatemanager==1.23.0
aws-cdk.aws-cloudformation==1.23.0
aws-cdk.aws-cloudwatch==1.23.0
aws-cdk.aws-codeguruprofiler==1.101.0
aws-cdk.aws-cognito==1.101.0
aws-cdk.aws-dynamodb==1.23.0
aws-cdk.aws-ec2==1.23.0
aws-cdk.aws-ecr==1.101.0
aws-cdk.aws-ecr-assets==1.101.0
aws-cdk.aws-efs==1.101.0
aws-cdk.aws-elasticloadbalancingv2==1.23.0
aws-cdk.aws-events==1.23.0
aws-cdk.aws-iam==1.23.0
aws-cdk.aws-kms==1.23.0
aws-cdk.aws-lambda==1.23.0
aws-cdk.aws-logs==1.23.0
aws-cdk.aws-route53==1.23.0
aws-cdk.aws-s3==1.23.0
aws-cdk.aws-s3-assets==1.23.0
aws-cdk.aws-signer==1.101.0
aws-cdk.aws-sns==1.23.0
aws-cdk.aws-sns-subscriptions==1.101.0
aws-cdk.aws-sqs==1.23.0
aws-cdk.aws-ssm==1.23.0
aws-cdk.cloud-assembly-schema==1.101.0
aws-cdk.core==1.23.0
aws-cdk.custom-resources==1.101.0
aws-cdk.cx-api==1.23.0
aws-cdk.region-info==1.23.0
boto3==1.17.62
botocore==1.20.62
cattrs==1.5.0
cdk-dynamo-table-viewer==3.1.2
# Editable Git install with no remote (cdk-workshop==0.0.1)
-e /Users/nurithofesh/PycharmProjects/cdk-workshop/cdk_workshop
constructs==3.3.75
importlib-metadata==4.0.1
iniconfig==1.1.1
jmespath==0.10.0
jsii==0.21.2
packaging==20.9
pluggy==0.13.1
publication==0.0.3
py==1.10.0
pyparsing==2.4.7
pytest==6.2.3
python-dateutil==2.8.1
s3transfer==0.4.2
six==1.15.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.4
zipp==3.4.1
